+++ 2_peripheral/2_uart_console/components/bsp/bsp_wrapper.h
@@ -1,0 +1,7 @@
+#ifndef BSP_WRAPPER_H_H_H
+#define BSP_WRAPPER_H_H_H
+
+#include "bsp_misc.h"
+#include "bsp_led.h"
+
+#endif