+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 - [BASE] [ADD]  增加工程模板
 - [PERIPHERAL] [ADD]  增加LED工程
+
+## 2022/04/08
+
+- [PERIPHERAL] [ADD]  ① 增加UART console工程 ② 增加UART select工程